chore: renames lib/widgets/round/ to lib/widgets/rounds/
