+++ frontend/lib/widgets/add_workout_form.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:wod_board_app/api.dart';
 import 'package:wod_board_app/models/workout.dart';
 import 'package:wod_board_app/widgets/misc/choice_list.dart';
-import 'package:wod_board_app/widgets/round/add_round.dart';
+import 'package:wod_board_app/widgets/rounds/add_round.dart';
 
 class AddWorkoutForm extends StatefulWidget {
   const AddWorkoutForm({super.key});
